Marginalise chi_var completely
There are clear consequences. While I have not checked it yet, it looks
like this model is closer to the hierarchical model where chi_var relies
on its prior value with the Normal distribution on chi in a given ROI.
This model fails to pool different ROIs _differently_. All ROIs
experience the same strength of regularisation disregarding heir
homogeneity. Unfortunately, often, this strength is insufficient for the
very homogenous ROIs. Maybe, still, it isn't something too relevant for
the real world susceptibility distributions.